pyproject.toml
[tool.poetry]
name = "repl-nix-stockinsighttracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.38.0"
yfinance = "^0.2.43"
plotly = "^5.24.0"
textblob = "^0.18.0.post0"
newsapi-python = "^0.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
